requirements.txt
python-dateutil
python-rclone
requests
selenium
selenium-wire
webdriver-manager
